Renamed GameInfoComponent to StolenRoofs
