+++ src/Components/InfoContainer/InfoContainer.jsx
@@ -5,7 +5,7 @@ import RoofStolen from '../RoofInfo/RoofStolen';
 import RoofNotFound from '../RoofInfo/RoofNotFound';
 import RefreshIndicator from 'material-ui/RefreshIndicator';
 import Leaderboard from '../Leaderboard/LeaderComponent';
-import StolenRoofs from '../GameInfoComponent/GameInfoComponent';
+import StolenRoofs from '../StolenRoofs/StolenRoofs';
 import Profile from '../Profile/ProfileComponent'
 import ChatComponent from '../ChatComponent/ChatComponent';
 import GameStatistics from '../GameStatistics/GameStatisticsComponent';
